working

+++ backend/management/src/main/java/com/library/management/repository/BookRepository.java
@@ -3,6 +3,6 @@ package com.library.management.repository;
 import com.library.management.entity.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepository extends JpaRepository<Book,Integer > {
+public interface BookRepository extends JpaRepository<Book, Long> {
+
 }
-
